cross_leverage_limit may decimals

+++ GateIo.Net/Objects/Models/GateIoPosition.cs
@@ -148,7 +148,7 @@ namespace GateIo.Net.Objects.Models
         /// Cross margin leverage
         /// </summary>
         [JsonPropertyName("cross_leverage_limit")]
-        public int? CrossLeverageLimit { get; set; }
+        public decimal? CrossLeverageLimit { get; set; }
         /// <summary>
         /// Update id
         /// </summary>
